agregado mas endpoints en recibo y usuario

+++ src/routes/usuario.routes.js
@@ -141,14 +141,30 @@ router.put('/:id', async (req, res) => {
 
 router.delete('/:id', async (req, res) => {
     const { id } = req.params;
-    
+    const { email, password } = req.body;
+
     try {
-        await prisma.usuario.delete({
-        where: {
-            usuarioId: parseInt(id)
+        if (!email || !password) {
+            return res.status(400).json({ error: 'Email and password are required' });
         }
+
+        const usuario = await prisma.usuario.findUnique({
+            where: {
+                email: email
+            }
         });
-    
+
+        const hashedPassword = crypto.createHash('sha256').update(password).digest('hex');
+        if (!usuario || usuario.usuarioId !== parseInt(id) || usuario.password !== hashedPassword) {
+            return res.status(401).json({ error: 'Invalid email, password, or user ID' });
+        }
+
+        await prisma.usuario.delete({
+            where: {
+                usuarioId: parseInt(id)
+            }
+        });
+
         res.json({ message: 'User deleted' });
     } catch (error) {
         console.error(error);
